+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 **Language**:
 
 - _Breaking_: Exclude `group`'s `by` columns from the partition. See #3490
+- _Breaking_: `lower` and `upper` are now in the `string` module and need to be
+  called via `string.lower` and `string.upper`. See #3913
 
 **Features**:
 
@@ -13,6 +15,10 @@
   `ceil`, `pi`, `exp`, `ln`, `log10`, `log`, `sqrt`, `degrees`, `radians`,
   `cos`, `acos`, `sin`, `asin`, `tan`, `atan` and `pow`.\
   Those functions are in the `math` module (@PrettyWood, #3909 & #3916)
+- Most standard string functions are now supported: `ltrim`, `rtrim`, `trim`,
+  `length`, `substring`, `replace`. Utility functions `starts_with`, `contains`
+  and `ends_with` are also available.\
+  Those functions are in the `string` module (@PrettyWood, #3913)
 
 **Fixes**: